separate liner list
