+++ Makefile
@@ -1,0 +1,2 @@
+stradermake: breakpoint_list.c strader.c 
+	gcc breakpoint_list.c strader.c -o strader
