+++ strader.c
@@ -1,3 +1,4 @@
+
 #include <stdio.h>
 #include <stdarg.h>
 #include <stdlib.h>
@@ -12,8 +13,7 @@
 #include <unistd.h>
 #include <errno.h>
 
-# define SUCCESSFUL 1
-# define FAILURE 0
+#include "breakpoint_list.h"
 
 //レジスタの値の色を変更するためのマクロ
 #define KNRM  "\x1B[0m"
@@ -25,100 +25,6 @@
 #define KCYN  "\x1B[36m"
 #define KWHT  "\x1B[37m"
 
-
-
-// ブレークポイントをヒープ領域に保存するためのリンクドリスト
-typedef unsigned long long int data_t;
-
-//構造体の作成
-typedef struct nodetag{
-
-    //ブレークポイントのアドレス
-    data_t address;
-
-    //ブレークポイントを設定する命令内容
-    data_t text;
-
-    //次の構造体へのアドレス
-    struct nodetag *next;
-}node_t;
-
-node_t *list = NULL;
-
-node_t *nodeNew(data_t dt, data_t dt2, node_t *nxt){
-    node_t * ndPtr;
-    ndPtr = malloc(sizeof(node_t));
-    if(ndPtr == NULL){
-        return NULL;
-    }else{
-        ndPtr -> address = dt;
-        ndPtr -> text = dt2;
-        return ndPtr;
-    }
-}
-
-int nodeAppend(node_t**ndPtrPtr, data_t address, data_t text){
-    node_t * ndPtr;
-    ndPtr = nodeNew(address, text, NULL);
-    if(ndPtr == NULL) return FAILURE;
-    while(*ndPtrPtr != NULL){
-        ndPtrPtr = &((*ndPtrPtr) -> next);
-    }
-    *ndPtrPtr = ndPtr;
-    return SUCCESSFUL;
-}
-
-// リストの中身を表示
-int listPrint(node_t *ndPtr){
-    //一番最後の構造体では次の構造体へのポインタにNULLが設定されているため
-    //NULLが出るまで繰り返す
-    while(ndPtr != NULL) {
-        printf("address :%016llx\ntext :%016llx\n", ndPtr -> address, ndPtr->text);
-        ndPtr = ndPtr -> next;
-    }
-}
-int node_findtext(node_t *ndPtr, data_t address){
-    int cnt;
-    cnt = 0;
-    while(ndPtr!= NULL){
-        
-        if(address == ndPtr ->address){
-            return ndPtr -> text;
-        }
-        ndPtr = ndPtr -> next;
-        cnt++;        
-    }
-    return FAILURE;
-}
-int node_findcnt(node_t *ndPtr, data_t address){
-    int cnt;
-    cnt = 0;
-    while(ndPtr!= NULL){
-        
-        if(address == ndPtr ->address){
-            return cnt;
-        }
-        ndPtr = ndPtr -> next;
-        cnt++;        
-    }
-    return FAILURE;
-}
-int nodeDelete(node_t **ndPtrPtr, int n){
-    node_t * ndPtr;
-    while (n>0 && *ndPtrPtr != NULL){
-        ndPtrPtr = &((*ndPtrPtr)->next);
-        n--;
-    }
-    if(*ndPtrPtr != NULL) {
-        ndPtr = (*ndPtrPtr) -> next;
-        free(*ndPtrPtr);
-        *ndPtrPtr = ndPtr;
-        return SUCCESSFUL;
-    }else{
-        return FAILURE;
-    }
-
-}
 
 //子プロセスの動作が完了し停止するまで待つ
 void p_wait(pid_t pid)
@@ -853,8 +759,8 @@ int main(int argc, char** argv){
         int intpid;
         // ここから始まる
 
-        printf("もしすでに動作しているプロセスをデバッグしたいときはプロセスIDを入力してください．\n");
-        printf("type \"q\" to quit.\n");
+        printf("すでに動作しているプロセスをデバッグするときはプロセスIDを入力.\n");
+        printf("type \"q\" to quit.\n > ");
         fgets(strpid ,sizeof(strpid) , stdin);
         if(!strcmp(strpid, "q\n")){
             exit(1);
